+++ components/layout.js
@@ -6,29 +6,43 @@ import ResponsiveAppBar from "../components/AppBar";
 import Breadcrumbs from "../components/Breadcrumbs";
 import * as React from 'react';
 import Typography from '@mui/material/Typography';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
 
 export const siteTitle = 'Haru TechLab';
 
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: "#303030",
+    },
+    secondary: {
+      main: "#e040fb",
+    },
+  },
+});
+
 export default function Layout({ children }) {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="../public/images/favicon.png" />
-      </Head>
-      <header className={styles.header}>
-        <ResponsiveAppBar />
-      </header>
-      <div>
-        <Breadcrumbs />
+    <ThemeProvider theme={theme}>
+      <div className={styles.container}>
+        <Head>
+          <title>Create Next App</title>
+          <meta name="description" content="Generated by create next app" />
+          <link rel="icon" href="../public/images/favicon.png" />
+        </Head>
+        <header className={styles.header}>
+          <ResponsiveAppBar color="primary" />
+        </header>
+        <div>
+          <Breadcrumbs />
+        </div>
+        <main className={styles.main}>{children}</main>
+        <footer className={styles.footer}>
+          <a>
+            ©2022 HARU
+          </a>
+        </footer>
       </div>
-      <main className={styles.main}>{children}</main>
-      <footer className={styles.footer}>
-        <a>
-          ©2022 HARU
-        </a>
-      </footer>
-    </div>
+    </ThemeProvider>
   )
 }
