+++ styles/globals.css
@@ -11,11 +11,6 @@ article {
   max-width: 960px;
 }
 
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
 * {
   box-sizing: border-box;
 }
